+++ src/Components/MapUploader.jsx
@@ -1,46 +1,33 @@
-import { MapContainer, TileLayer, GeoJSON } from "react-leaflet";
-import { useState } from "react";
+import { MapContainer, TileLayer, GeoJSON, useMap } from "react-leaflet";
+import { useState, useEffect } from "react";
 import { DOMParser as Dom } from "xmldom";
 import * as toGeoJSON from "@tmcw/togeojson";
 import { FaFolderOpen } from "react-icons/fa";
-import { Box, IconButton, useMediaQuery } from "@mui/material";
-import { useTheme } from "@mui/material/styles";
-import { useMap } from "react-leaflet";
-import React, { useEffect } from "react";
+import { Box, IconButton } from "@mui/material";
+import useResponsive from "../Hooks/Responsive";
+import { containerBox, iconButtonBase } from "../Styles/MapStyling";
 import "leaflet/dist/leaflet.css";
+
+const ResizeHandler = () => {
+  const map = useMap();
+  useEffect(() => {
+    const handleResize = () => map.invalidateSize();
+    window.addEventListener("resize", handleResize);
+    return () => window.removeEventListener("resize", handleResize);
+  }, [map]);
+
+  return null;
+};
 
 const MapUploaderKML = () => {
   const [geoData, setGeoData] = useState(null);
-  const theme = useTheme();
-  const isSmallScreen = useMediaQuery(theme.breakpoints.down("sm"));
-  const isTablet = useMediaQuery(theme.breakpoints.between("sm", "md"));
-  const isDesktop = useMediaQuery(theme.breakpoints.up("lg"));
+  const { isSmallScreen, isTablet, isDesktop } = useResponsive();
 
   let marginLeftValue;
-
-  if (isSmallScreen) {
-    marginLeftValue = "0";
-  } else if (isTablet) {
-    marginLeftValue = "-5px";
-  } else if (isDesktop) {
-    marginLeftValue = "-10px";
-  } else {
-    marginLeftValue = "-7px";
-  }
-  const ResizeHandler = () => {
-    const map = useMap();
-
-    useEffect(() => {
-      const handleResize = () => {
-        map.invalidateSize();
-      };
-
-      window.addEventListener("resize", handleResize);
-      return () => window.removeEventListener("resize", handleResize);
-    }, [map]);
-
-    return null;
-  };
+  if (isSmallScreen) marginLeftValue = "0";
+  else if (isTablet) marginLeftValue = "-5px";
+  else if (isDesktop) marginLeftValue = "-10px";
+  else marginLeftValue = "-7px";
 
   const handleFileUpload = (event) => {
     const file = event.target.files[0];
@@ -48,8 +35,7 @@ const MapUploaderKML = () => {
       const reader = new FileReader();
       reader.onload = (e) => {
         try {
-          const kmlText = e.target.result;
-          const kmlDom = new Dom().parseFromString(kmlText, "text/xml");
+          const kmlDom = new Dom().parseFromString(e.target.result, "text/xml");
           const converted = toGeoJSON.kml(kmlDom);
           setGeoData(converted);
         } catch (err) {
@@ -70,37 +56,9 @@ const MapUploaderKML = () => {
       alignItems="center"
       width="100%"
       mt={4}
-      position="relative"
     >
-      <Box
-        sx={{
-          width: {
-            xs: "95vw",
-            sm: "90vw",
-            md: "80vw",
-            lg: "70vw",
-          },
-          height: {
-            xs: "70vh",
-            sm: "80vh",
-            md: "85vh",
-            lg: "90vh",
-          },
-          position: "relative",
-        }}
-      >
-        <IconButton
-          component="label"
-          sx={{
-            position: "absolute",
-            top: 78,
-            left: 16,
-            zIndex: 1000,
-            ml: marginLeftValue,
-            backgroundColor: "white",
-            "&:hover": { backgroundColor: "#f0f0f0" },
-          }}
-        >
+      <Box sx={containerBox}>
+        <IconButton component="label" sx={iconButtonBase(marginLeftValue)}>
           <FaFolderOpen size={24} color="#0000FF" />
           <input type="file" accept=".kml" onChange={handleFileUpload} hidden />
         </IconButton>
