+++ src/Hooks/Responsive.js
@@ -1,0 +1,14 @@
+// src/hooks/useResponsive.js
+import { useTheme } from "@mui/material/styles";
+import useMediaQuery from "@mui/material/useMediaQuery";
+
+const useResponsive = () => {
+  const theme = useTheme();
+  const isSmallScreen = useMediaQuery(theme.breakpoints.down("sm"));
+  const isTablet = useMediaQuery(theme.breakpoints.between("sm", "md"));
+  const isDesktop = useMediaQuery(theme.breakpoints.up("lg"));
+
+  return { isSmallScreen, isTablet, isDesktop };
+};
+
+export default useResponsive;
